requirements.txt
asgiref==3.9.1
Django==4.2.23
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.1
Markdown==3.8.2
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
